Implement TodoItem component
